+++ attributes/ngx_pagespeed.rb
@@ -1,13 +1,13 @@
 # # # # # # # # #
 # ngx_pagespeed #
 # # # # # # # # #
-default['nginx']['ngx_pagespeed']['version'] = '1.9.32.2-beta'
+default['nginx']['ngx_pagespeed']['version'] = '1.9.32.6-beta'
 default['nginx']['ngx_pagespeed']['url'] = "https://codeload.github.com/pagespeed/ngx_pagespeed/tar.gz/v#{node['nginx']['ngx_pagespeed']['version']}"
-default['nginx']['ngx_pagespeed']['checksum'] = 'c728c2cd5f93af183f53ba02164442ddfc552e146f953a9110a6e7fa7e0fb1be'
+default['nginx']['ngx_pagespeed']['checksum'] = '281884005ce4438913fe8fdc114a6844071a87d3189a5ce0d5d39bbbd785b9fb'
 
-default['nginx']['ngx_pagespeed']['psol']['version'] = '1.9.32.2'
+default['nginx']['ngx_pagespeed']['psol']['version'] = '1.9.32.6'
 default['nginx']['ngx_pagespeed']['psol']['url'] = "https://dl.google.com/dl/page-speed/psol/#{node['nginx']['ngx_pagespeed']['psol']['version']}.tar.gz"
-default['nginx']['ngx_pagespeed']['checksum'] = 'd161bbad350bf2242cf4d7e179b5244e1f4eed767c0e67c5c7109f7eb08bf224'
+default['nginx']['ngx_pagespeed']['checksum'] = 'b2285a1a9f03a5ff4107852480edf28fedb1804ab1dfde45ef0abb4150ef9912'
 
 default['nginx']['ngx_pagespeed']['cache-url'] = '/var/ngx_pagespeed_cache'
 default['nginx']['ngx_pagespeed']['cache-size'] = '512M'
@@ -15,8 +15,8 @@ default['nginx']['ngx_pagespeed']['enabled'] = 'on'
 default['nginx']['ngx_pagespeed']['canonicalize_javascript_libraries'] = false
 default['nginx']['ngx_pagespeed']['FileCachePath'] = node['nginx']['ngx_pagespeed']['cache-url']
 default['nginx']['ngx_pagespeed']['CacheFlushFilename'] = 'cache.flush'
-default['nginx']['ngx_pagespeed']['EnableFilters'] = 'collapse_whitespace,combine_css,combine_javascript,remove_comments,sprite_images,extend_cache'
-default['nginx']['ngx_pagespeed']['DisableFilters'] = 'defer_javascript'
+default['nginx']['ngx_pagespeed']['EnableFilters'] = 'convert_gif_to_png,insert_image_dimensions,lazyload_images,flatten_css_imports,insert_dns_prefetch,collapse_whitespace,combine_css,combine_javascript,sprite_images,recompress_images,rewrite_images,extend_cache,make_google_analytics_async'
+default['nginx']['ngx_pagespeed']['DisableFilters'] = 'defer_javascript,remove_comments'
 
 default['nginx']['ngx_pagespeed']['Statistics'] = 'on'
 default['nginx']['ngx_pagespeed']['StatisticsLogging'] = 'on'
@@ -41,3 +41,4 @@ default['nginx']['ngx_pagespeed']['CacheFlushPollIntervalSec'] = 600_000
 default['nginx']['ngx_pagespeed']['DownstreamCachePurgeMethod'] = 'PURGE'
 default['nginx']['ngx_pagespeed']['DownstreamCacheRewrittenPercentageThreshold'] = 95
 default['nginx']['ngx_pagespeed']['RetainComment'] = 'esi*'
+default['nginx']['ngx_pagespeed']['Disallow'] = ''